Extract interaction with LocalServerDomain
